split tests into separate files per class type
