--- cli/release_notes/tests/test_release_notes.py
@@ -1,1009 +1,0 @@
-import httplib
-import os
-import shutil
-import tempfile
-import unittest
-
-from datetime import datetime
-from datetime import timedelta
-
-import requests
-import responses
-
-from release_notes.generator import BaseReleaseNotesGenerator
-from release_notes.generator import StaticReleaseNotesGenerator
-from release_notes.generator import DirectoryReleaseNotesGenerator
-from release_notes.generator import GithubReleaseNotesGenerator
-
-from release_notes.parser import BaseChangeNotesParser
-from release_notes.parser import ChangeNotesLinesParser
-from release_notes.parser import CommentingGithubIssuesParser
-from release_notes.parser import GithubIssuesParser
-from release_notes.parser import IssuesParser
-
-from release_notes.provider import BaseChangeNotesProvider
-from release_notes.provider import StaticChangeNotesProvider
-from release_notes.provider import DirectoryChangeNotesProvider
-from release_notes.provider import GithubChangeNotesProvider
-
-from release_notes.exceptions import GithubApiNotFoundError
-from release_notes.exceptions import LastReleaseTagNotFoundError
-
-from release_notes.tests.util_github_api import GithubApiTestMixin
-
-__location__ = os.path.realpath(
-    os.path.join(os.getcwd(), os.path.dirname(__file__)))
-
-
-class DummyParser(BaseChangeNotesParser):
-
-    def parse(self, change_note):
-        pass
-
-    def _render(self):
-        return 'dummy parser output'.format(self.title)
-
-
-class TestBaseReleaseNotesGenerator(unittest.TestCase):
-
-    def test_render_no_parsers(self):
-        release_notes = BaseReleaseNotesGenerator()
-        content = release_notes.render()
-        self.assertEqual(content, '')
-
-    def test_render_dummy_parsers(self):
-        release_notes = BaseReleaseNotesGenerator()
-        release_notes.parsers.append(DummyParser('Dummy 1'))
-        release_notes.parsers.append(DummyParser('Dummy 2'))
-        expected = u'# Dummy 1\r\n\r\ndummy parser output\r\n\r\n' +\
-                   u'# Dummy 2\r\n\r\ndummy parser output'
-        self.assertEqual(release_notes.render(), expected)
-
-
-class TestStaticReleaseNotesGenerator(unittest.TestCase):
-
-    def test_init_parser(self):
-        release_notes = StaticReleaseNotesGenerator([])
-        assert len(release_notes.parsers) == 3
-
-
-class TestDirectoryReleaseNotesGenerator(unittest.TestCase):
-
-    def test_init_parser(self):
-        release_notes = DirectoryReleaseNotesGenerator('change_notes')
-        assert len(release_notes.parsers) == 3
-
-    def test_full_content(self):
-        change_notes_dir = os.path.join(
-            __location__,
-            'change_notes',
-            'full',
-        )
-        release_notes = DirectoryReleaseNotesGenerator(
-            change_notes_dir,
-        )
-        
-        content = release_notes()
-        expected = "# Critical Changes\r\n\r\n* This will break everything!\r\n\r\n# Changes\r\n\r\nHere's something I did. It was really cool\r\nOh yeah I did something else too!\r\n\r\n# Issues Closed\r\n\r\n#2345\r\n#6236"
-        print expected
-        print '-------------------------------------'
-        print content
-
-        self.assertEquals(content, expected) 
-
-
-class TestBaseChangeNotesParser(unittest.TestCase):
-    pass
-
-
-class TestChangeNotesLinesParser(unittest.TestCase):
-
-    def setUp(self):
-        self.title = 'Title'
-
-    def test_parse_no_start_line(self):
-        change_note = 'foo\r\nbar\r\n'
-        parser = ChangeNotesLinesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [])
-
-    def test_parse_start_line_no_content(self):
-        change_note = '# {}\r\n\r\n'.format(self.title)
-        parser = ChangeNotesLinesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [])
-
-    def test_parse_start_line_no_end_line(self):
-        change_note = '# {}\r\nfoo\r\nbar'.format(self.title)
-        parser = ChangeNotesLinesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, ['foo', 'bar'])
-
-    def test_parse_start_line_end_at_header(self):
-        change_note = '# {}\r\nfoo\r\n# Another Header\r\nbar'.format(self.title)
-        parser = ChangeNotesLinesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, ['foo'])
-
-    def test_parse_start_line_no_content_no_end_line(self):
-        change_note = '# {}'.format(self.title)
-        parser = ChangeNotesLinesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [])
-
-    def test_parse_multiple_start_lines_without_end_lines(self):
-        change_note = '# {0}\r\nfoo\r\n# {0}\r\nbar\r\n'.format(self.title)
-        parser = ChangeNotesLinesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, ['foo', 'bar'])
-
-    def test_parse_multiple_start_lines_with_end_lines(self):
-        change_note = '# {0}\r\nfoo\r\n\r\n# {0}\r\nbar\r\n\r\nincluded\r\n\r\n# not included'.format(self.title)
-        parser = ChangeNotesLinesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, ['foo', 'bar', 'included'])
-
-    def test_render_no_content(self):
-        parser = ChangeNotesLinesParser(None, self.title)
-        self.assertEqual(parser.render(), None)
-
-    def test_render_one_content(self):
-        parser = ChangeNotesLinesParser(None, self.title)
-        content = ['foo']
-        parser.content = content
-        self.assertEqual(parser.render(),
-                         '# {}\r\n\r\n{}'.format(self.title, content[0]))
-
-    def test_render_multiple_content(self):
-        parser = ChangeNotesLinesParser(None, self.title)
-        content = ['foo', 'bar']
-        parser.content = content
-        self.assertEqual(parser.render(),
-                         '# {}\r\n\r\n{}'.format(self.title, '\r\n'.join(content)))
-
-class TestIssuesParser(unittest.TestCase):
-
-    def setUp(self):
-        self.title = 'Issues'
-
-    def test_issue_numbers(self):
-        change_note = '# {}\r\nfix #2\r\nfix #3\r\nfix #5\r\n'.format(self.title)
-        parser = IssuesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [2, 3, 5])
-
-    def test_issue_numbers_and_other_numbers(self):
-        change_note = '# {}\r\nfixes #2 but not # 3 or 5'.format(self.title)
-        parser = IssuesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [2])
-
-    def test_multiple_issue_numbers_per_line(self):
-        change_note = '# {}\r\nfix #2 also does fix #3 and fix #5\r\n'.format(
-            self.title)
-        parser = IssuesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [2, 3, 5])
-
-class TestGithubIssuesParser(unittest.TestCase, GithubApiTestMixin):
-
-    def setUp(self):
-        self.init_github()
-
-        self.title = 'Issues'
-        # Set up the mock release_tag lookup response
-        self.issue_number_valid = 123
-        self.issue_number_invalid = 456
-
-    def test_issue_numbers(self):
-        change_note = '# {}\r\nFixes #2, Closed #3 and Resolve #5'.format(
-            self.title)
-        parser = GithubIssuesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [2, 3, 5])
-
-    def test_issue_numbers_and_other_numbers(self):
-        change_note = '# {}\r\nFixes #2 but not #5'.format(
-            self.title)
-        parser = GithubIssuesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [2])
-
-    def test_no_issue_numbers(self):
-        change_note = '# {}\r\n#2 and #3 are fixed by this change'.format(
-            self.title)
-        parser = GithubIssuesParser(None, self.title)
-        parser.parse(change_note)
-        self.assertEqual(parser.content, [])
-
-    @responses.activate
-    def test_render_issue_number_valid(self):
-        api_url = '{}/issues/{}'.format(
-            self.repo_api_url, self.issue_number_valid)
-        expected_response = self._get_expected_issue(self.issue_number_valid)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response,
-        )
-        generator = self._create_generator()
-        parser = GithubIssuesParser(generator, self.title)
-        parser.content = [self.issue_number_valid]
-        expected_render = '# {}\r\n\r\n#{}: {}'.format(
-            self.title,
-            self.issue_number_valid,
-            expected_response['title'],
-        )
-        self.assertEqual(parser.render(), expected_render)
-
-    @responses.activate
-    def test_render_issue_number_invalid(self):
-        api_url = '{}/issues/{}'.format(
-            self.repo_api_url, self.issue_number_invalid)
-        expected_response = {
-            'message': 'Not Found',
-            'documentation_url': 'https://developer.github.com/v3',
-        }
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response,
-            status=httplib.NOT_FOUND,
-        )
-        generator = self._create_generator()
-        parser = GithubIssuesParser(generator, self.title)
-        parser.content = [self.issue_number_invalid]
-        with self.assertRaises(GithubApiNotFoundError):
-            parser.render()
-
-    def _create_generator(self):
-        generator = GithubReleaseNotesGenerator(self.github_info.copy(), 'prod/1.1')
-        return generator
-
-
-class TestCommentingGithubIssuesParser(unittest.TestCase, GithubApiTestMixin):
-
-    def setUp(self):
-        self.init_github()
-
-        self.title = 'Issues'
-        self.issue_number_without_comments = 1
-        self.issue_number_with_beta_comment = 2
-        self.issue_number_without_beta_comment = 3
-        self.issue_number_with_prod_comment = 4
-        self.issue_number_without_prod_comment = 5
-        self.tag_prod = 'prod/1.2'
-        self.tag_beta = 'beta/1.2-Beta_3'
-        self.tag_not_prod_or_beta = 'foo'
-        self.version_number_prod = '1.1'
-        self.version_number_beta = '1.2 (Beta 3)'
-
-    def _create_generator(self, tag):
-        generator = GithubReleaseNotesGenerator(self.github_info.copy(), tag)
-        return generator
-    
-    @responses.activate
-    def test_render_issue_without_comments(self):
-        issue_number = self.issue_number_without_comments
-        tag = self.tag_not_prod_or_beta
-        
-        # Mock the issue
-        api_url = '{}/issues/{}'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_issue = self._get_expected_issue(issue_number)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_issue,
-        )
-
-        # Mock the comments list
-        api_url = '{}/issues/{}/comments'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            body=[],
-            content_type='application/json',   
-        )
-
-        generator = self._create_generator(tag)
-        parser = CommentingGithubIssuesParser(generator, self.title)
-        parser.content = [issue_number]
-        expected_render = '# {}\r\n\r\n#{}: {}'.format(
-            self.title,
-            issue_number,
-            expected_issue['title'],
-        )
-        self.assertEqual(parser.render(), expected_render)
-
-        # Only 2 api calls were made, ensuring comment creation
-        # was not attempted
-        self.assertEqual(len(responses.calls._calls), 2)
-       
-
-    @responses.activate 
-    def test_render_issue_with_beta_comment(self):
-        issue_number = self.issue_number_with_beta_comment
-        tag = self.tag_beta
-        
-        # Mock the issue
-        api_url = '{}/issues/{}'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_issue = self._get_expected_issue(issue_number)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_issue,
-        )
-
-        # Mock the comments list
-        api_url = '{}/issues/{}/comments'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_comment_1 = self._get_expected_issue_comment(
-            CommentingGithubIssuesParser.message_beta,
-        )
-        expected_comments = [
-            expected_comment_1,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_comments,
-        )
-
-        generator = self._create_generator(tag)
-        parser = CommentingGithubIssuesParser(generator, self.title)
-        parser.content = [issue_number]
-        expected_render = '# {}\r\n\r\n#{}: {}'.format(
-            self.title,
-            issue_number,
-            expected_issue['title'],
-        )
-        self.assertEqual(parser.render(), expected_render)
-
-        # Only 2 api calls were made, ensuring comment creation
-        # was not attempted
-        self.assertEqual(len(responses.calls._calls), 2)
-
-    @responses.activate
-    def test_render_issue_without_beta_comment(self):
-        issue_number = self.issue_number_without_beta_comment
-        tag = self.tag_beta
-        
-        # Mock the issue
-        api_url = '{}/issues/{}'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_issue = self._get_expected_issue(issue_number)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_issue,
-        )
-
-        # Mock the comments list
-        api_url = '{}/issues/{}/comments'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_comment_1 = self._get_expected_issue_comment(
-            'Some other comment',
-        )
-        expected_comments = [
-            expected_comment_1,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            body=[],
-            content_type='application/json',   
-        )
-
-        # Mock the comment post response
-        api_url = '{}/issues/{}/comments'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_comment_1 = self._get_expected_issue_comment(
-            '{} {}'.format(
-                CommentingGithubIssuesParser.message_beta,
-                self.version_number_beta,        
-            )
-        )
-        responses.add(
-            method=responses.POST,
-            url=api_url,
-            json=expected_comment_1,
-        )
-
-        generator = self._create_generator(tag)
-        parser = CommentingGithubIssuesParser(generator, self.title)
-        parser.content = [issue_number]
-        expected_render = '# {}\r\n\r\n#{}: {}'.format(
-            self.title,
-            issue_number,
-            expected_issue['title'],
-        )
-        self.assertEqual(parser.render(), expected_render)
-
-        # 3 api calls were made, ensuring comment creation
-        # was attempted
-        self.assertEqual(len(responses.calls._calls), 3)
-
-    @responses.activate 
-    def test_render_issue_with_prod_comment(self):
-        issue_number = self.issue_number_with_prod_comment
-        tag = self.tag_prod
-        
-        # Mock the issue
-        api_url = '{}/issues/{}'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_issue = self._get_expected_issue(issue_number)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_issue,
-        )
-
-        # Mock the comments list
-        api_url = '{}/issues/{}/comments'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_comment_1 = self._get_expected_issue_comment(
-            CommentingGithubIssuesParser.message_prod,
-        )
-        expected_comments = [
-            expected_comment_1,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_comments,
-        )
-
-        generator = self._create_generator(tag)
-        parser = CommentingGithubIssuesParser(generator, self.title)
-        parser.content = [issue_number]
-        expected_render = '# {}\r\n\r\n#{}: {}'.format(
-            self.title,
-            issue_number,
-            expected_issue['title'],
-        )
-        self.assertEqual(parser.render(), expected_render)
-
-        # Only 2 api calls were made, ensuring comment creation
-        # was not attempted
-        self.assertEqual(len(responses.calls._calls), 2)
-
-    @responses.activate
-    def test_render_issue_without_prod_comment(self):
-        issue_number = self.issue_number_without_prod_comment
-        tag = self.tag_prod
-        
-        # Mock the issue
-        api_url = '{}/issues/{}'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_issue = self._get_expected_issue(issue_number)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_issue,
-        )
-
-        # Mock the comments list
-        api_url = '{}/issues/{}/comments'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_comment_1 = self._get_expected_issue_comment(
-            'Some other comment',
-        )
-        expected_comments = [
-            expected_comment_1,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            body=[],
-            content_type='application/json',   
-        )
-
-        # Mock the comment post response
-        api_url = '{}/issues/{}/comments'.format(
-            self.repo_api_url,
-            issue_number,
-        )
-        expected_comment_1 = self._get_expected_issue_comment(
-            '{} {}'.format(
-                CommentingGithubIssuesParser.message_prod,
-                self.version_number_prod,        
-            )
-        )
-        responses.add(
-            method=responses.POST,
-            url=api_url,
-            json=expected_comment_1,
-        )
-
-        generator = self._create_generator(tag)
-        parser = CommentingGithubIssuesParser(generator, self.title)
-        parser.content = [issue_number]
-        expected_render = '# {}\r\n\r\n#{}: {}'.format(
-            self.title,
-            issue_number,
-            expected_issue['title'],
-        )
-        self.assertEqual(parser.render(), expected_render)
-
-        # 3 api calls were made, ensuring comment creation
-        # was attempted
-        self.assertEqual(len(responses.calls._calls), 3)
-
-class TestBaseChangeNotesProvider(unittest.TestCase):
-
-    def test_init(self):
-        provider = BaseChangeNotesProvider('test')
-        assert provider.release_notes_generator == 'test'
-
-    def test_call_raises_notimplemented(self):
-        provider = BaseChangeNotesProvider('test')
-        self.assertRaises(NotImplementedError, provider.__call__)
-
-
-class TestStaticChangeNotesProvider(unittest.TestCase):
-
-    def test_empty_list(self):
-        provider = StaticChangeNotesProvider('test', [])
-        assert list(provider()) == []
-
-    def test_single_item_list(self):
-        provider = StaticChangeNotesProvider('test', ['abc'])
-        assert list(provider()) == ['abc']
-
-    def test_multi_item_list(self):
-        provider = StaticChangeNotesProvider('test', ['abc', 'd', 'e'])
-        assert list(provider()) == ['abc', 'd', 'e']
-
-
-class TestDirectoryChangeNotesProvider(unittest.TestCase):
-
-    def get_empty_dir(self):
-        tempdir = tempfile.mkdtemp()
-        return os.path.join(tempdir)
-
-    def get_dir_content(self, path):
-        dir_content = []
-        for item in os.listdir(path):
-            item_path = '{}/{}'.format(path, item)
-            dir_content.append(open(item_path, 'r').read())
-        return dir_content
-
-    def test_empty_directory(self):
-        directory = self.get_empty_dir()
-        provider = DirectoryChangeNotesProvider('test', directory)
-        dir_content = self.get_dir_content(directory)
-        assert list(provider()) == dir_content
-        shutil.rmtree(directory)
-
-    def test_single_item_directory(self):
-        directory = '{}/change_notes/single/'.format(__location__)
-        provider = DirectoryChangeNotesProvider('test', directory)
-        dir_content = self.get_dir_content(directory)
-        assert list(provider()) == dir_content
-
-    def test_multi_item_directory(self):
-        directory = '{}/change_notes/multi/'.format(__location__)
-        provider = DirectoryChangeNotesProvider('test', directory)
-        dir_content = self.get_dir_content(directory)
-        assert list(provider()) == dir_content
-
-
-class TestGithubChangeNotesProvider(unittest.TestCase, GithubApiTestMixin):
-
-    def setUp(self):
-        # Set up the mock release_tag lookup response
-        self.repo_api_url = 'https://api.github.com/repos/TestOwner/TestRepo'
-        # Tag that does not exist
-        self.invalid_tag = 'prod/1.4'
-        # The current production release
-        self.current_tag = 'prod/1.3'
-        # The previous production release with no change notes vs 1.3
-        self.last_tag = 'prod/1.2'
-        # The second previous production release with one change note vs 1.3
-        self.last2_tag = 'prod/1.1'
-        # The third previous production release with three change notes vs 1.3
-        self.last3_tag = 'prod/1.0'
-
-        self.current_tag_sha = self._random_sha()
-        self.last_tag_sha = self._random_sha()
-
-        self.github_info = {
-            'github_owner': 'TestOwner',
-            'github_repo': 'TestRepo',
-            'github_username': 'TestUser',
-            'github_password': 'TestPass',
-        }
-
-    def _create_generator(self, current_tag, last_tag=None):
-        generator = GithubReleaseNotesGenerator(
-            self.github_info.copy(),
-            current_tag,
-            last_tag=last_tag
-        )
-        return generator
-
-    @responses.activate
-    def test_invalid_current_tag(self):
-        api_url = '{}/git/refs/tags/{}'.format(
-            self.repo_api_url, self.invalid_tag)
-        expected_response = {
-            'message': 'Not Found',
-            'documentation_url': 'https://developer.github.com/v3'
-        }
-
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response,
-            status=httplib.NOT_FOUND,
-        )
-
-        generator = self._create_generator(self.invalid_tag)
-        provider = GithubChangeNotesProvider(generator, self.invalid_tag)
-        with self.assertRaises(GithubApiNotFoundError):
-            provider.current_tag_info
-
-    @responses.activate
-    def test_current_tag_without_last(self):
-
-        # Mock the current tag ref
-        api_url = '{}/git/refs/tags/{}'.format(
-            self.repo_api_url, self.current_tag)
-        expected_response_current_tag_ref = self._get_expected_tag_ref(
-            self.current_tag, self.current_tag_sha)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_current_tag_ref,
-        )
-
-        # Mock the current tag
-        api_url = '{}/git/tags/{}'.format(self.repo_api_url,
-                                          self.current_tag_sha)
-        expected_response_current_tag = self._get_expected_tag(
-            self.current_tag, self.current_tag_sha)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_current_tag,
-        )
-
-        # Mock the last tag ref
-        api_url = '{}/git/refs/tags/{}'.format(
-            self.repo_api_url, self.last_tag)
-        expected_response_last_tag_ref = self._get_expected_tag_ref(
-            self.last_tag, self.last_tag_sha)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_last_tag_ref,
-        )
-
-        # Mock the last tag
-        api_url = '{}/git/tags/{}'.format(self.repo_api_url, self.last_tag_sha)
-        expected_response_last_tag = self._get_expected_tag(
-            self.last_tag, self.last_tag_sha)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_last_tag,
-        )
-
-        # Mock the list all tags call
-        api_url = '{}/git/refs/tags/prod/'.format(self.repo_api_url)
-        expected_response_list_tag_refs = [
-            expected_response_current_tag_ref,
-            expected_response_last_tag_ref,
-            self._get_expected_tag_ref(self.last2_tag, 'last2_tag_sha'),
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_list_tag_refs,
-        )
-
-        generator = self._create_generator(self.current_tag)
-        provider = GithubChangeNotesProvider(generator, self.current_tag)
-
-        self.assertEquals(provider.current_tag_info[
-                          'ref'], expected_response_current_tag_ref)
-        self.assertEquals(provider.current_tag_info[
-                          'tag'], expected_response_current_tag)
-        self.assertEquals(provider.last_tag_info[
-                          'ref'], expected_response_last_tag_ref)
-        self.assertEquals(provider.last_tag_info[
-                          'tag'], expected_response_last_tag)
-
-    @responses.activate
-    def test_current_tag_without_last_no_last_found(self):
-        # Mock the current tag ref
-        api_url = '{}/git/refs/tags/{}'.format(
-            self.repo_api_url, self.current_tag)
-        expected_response_current_tag_ref = self._get_expected_tag_ref(
-            self.current_tag, self.current_tag_sha)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_current_tag_ref,
-        )
-
-        # Mock the current tag
-        api_url = '{}/git/tags/{}'.format(self.repo_api_url,
-                                          self.current_tag_sha)
-        expected_response_current_tag = self._get_expected_tag(
-            self.current_tag, self.current_tag_sha)
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_current_tag,
-        )
-
-        # Mock the list all tags call
-        api_url = '{}/git/refs/tags/prod/'.format(self.repo_api_url)
-        expected_response_list_tag_refs = [
-            expected_response_current_tag_ref,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_list_tag_refs,
-        )
-        generator = self._create_generator(self.current_tag)
-        provider = GithubChangeNotesProvider(generator, self.current_tag)
-
-        with self.assertRaises(LastReleaseTagNotFoundError):
-            provider.last_tag
-
-        with self.assertRaises(LastReleaseTagNotFoundError):
-            provider.last_tag_info
-
-    def _mock_pull_request_tags(self):
-        # Mock the current tag ref
-        api_url = '{}/git/refs/tags/{}'.format(
-            self.repo_api_url, 
-            self.current_tag,
-        )
-        expected_response_current_tag_ref = self._get_expected_tag_ref(
-            self.current_tag, 
-            self.current_tag_sha,
-        )
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_current_tag_ref,
-        )
-
-        # Mock the current tag
-        api_url = '{}/git/tags/{}'.format(
-            self.repo_api_url,
-            self.current_tag_sha,
-        )
-        expected_response_current_tag = self._get_expected_tag(
-            self.current_tag,
-            self.current_tag_sha,
-            datetime.now(),
-        )
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_current_tag,
-        )
-
-        # Mock the last tag ref
-        api_url = '{}/git/refs/tags/{}'.format(
-            self.repo_api_url, 
-            self.last_tag
-        )
-        expected_response_last_tag_ref = self._get_expected_tag_ref(
-            self.last_tag, 
-            self.last_tag_sha,
-        )
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_last_tag_ref,
-        )
-
-        # Mock the last tag
-        api_url = '{}/git/tags/{}'.format(
-            self.repo_api_url, 
-            self.last_tag_sha,
-        )
-        expected_response_last_tag = self._get_expected_tag(
-            self.last_tag, 
-            self.last_tag_sha,
-            datetime.now() - timedelta(days=1),
-        )
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_last_tag,
-        )
-
-    @responses.activate
-    def test_no_pull_requests_in_repo(self):
-        # Mock the tag calls
-        self._mock_pull_request_tags()
-
-        # Mock the list all pull requests call
-        api_url = '{}/pulls'.format(self.repo_api_url)
-        expected_response_list_pull_requests = []
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            body=expected_response_list_pull_requests,
-            content_type='application/json',
-        )
-
-        generator = self._create_generator(self.current_tag, self.last_tag)
-        provider = GithubChangeNotesProvider(generator, self.current_tag, self.last_tag)
-        self.assertEquals(list(provider()), [])
-
-    @responses.activate
-    def test_no_pull_requests_in_range(self):
-        # Mock the tag calls
-        self._mock_pull_request_tags()
-
-        # Mock the list all pull requests call
-        api_url = '{}/pulls'.format(self.repo_api_url)
-        expected_pull_request_1 = self._get_expected_pull_request(
-            pull_id=1,
-            issue_number=101,
-            body='pull 1',
-            merged_date=datetime.now() - timedelta(days=2),
-        )
-        expected_response_list_pull_requests = [
-            expected_pull_request_1,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_list_pull_requests,
-        )
-
-        generator = self._create_generator(self.current_tag, self.last_tag)
-        provider = GithubChangeNotesProvider(generator, self.current_tag, self.last_tag)
-        self.assertEquals(list(provider()), [])
-
-    @responses.activate
-    def test_one_pull_request_in_range(self):
-        # Mock the tag calls
-        self._mock_pull_request_tags()
-
-        # Mock the list all pull requests call
-        api_url = '{}/pulls'.format(self.repo_api_url)
-        expected_pull_request_1 = self._get_expected_pull_request(
-            pull_id=1,
-            issue_number=101,
-            body='pull 1',
-            merged_date=datetime.now() - timedelta(seconds=60),
-        )
-        expected_pull_request_2 = self._get_expected_pull_request(
-            pull_id=2,
-            issue_number=102,
-            body='pull 2',
-            merged_date=datetime.now() - timedelta(days=2),
-        )
-        expected_response_list_pull_requests = [
-            expected_pull_request_1,
-            expected_pull_request_2,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_list_pull_requests,
-        )
-
-        generator = self._create_generator(self.current_tag, self.last_tag)
-        provider = GithubChangeNotesProvider(generator, self.current_tag, self.last_tag)
-        self.assertEquals(list(provider()), ['pull 1'])
-
-    @responses.activate
-    def test_multiple_pull_requests_in_range(self):
-        # Mock the tag calls
-        self._mock_pull_request_tags()
-
-        # Mock the list all pull requests call
-        api_url = '{}/pulls'.format(self.repo_api_url)
-        expected_pull_request_1 = self._get_expected_pull_request(
-            pull_id=1,
-            issue_number=101,
-            body='pull 1',
-            merged_date=datetime.now() - timedelta(seconds=60),
-        )
-        expected_pull_request_2 = self._get_expected_pull_request(
-            pull_id=2,
-            issue_number=102,
-            body='pull 2',
-            merged_date=datetime.now() - timedelta(seconds=90),
-        )
-        expected_pull_request_3 = self._get_expected_pull_request(
-            pull_id=3,
-            issue_number=103,
-            body='pull 3',
-            merged_date=datetime.now() - timedelta(seconds=120),
-        )
-        expected_pull_request_4 = self._get_expected_pull_request(
-            pull_id=4,
-            issue_number=104,
-            body='pull 4',
-            merged_date=datetime.now() - timedelta(days=4),
-        )
-        expected_pull_request_5 = self._get_expected_pull_request(
-            pull_id=5,
-            issue_number=105,
-            body='pull 5',
-            merged_date=datetime.now() - timedelta(days=5),
-        )
-        expected_response_list_pull_requests = [
-            expected_pull_request_1,
-            expected_pull_request_2,
-            expected_pull_request_3,
-            expected_pull_request_4,
-            expected_pull_request_5,
-        ]
-        responses.add(
-            method=responses.GET,
-            url=api_url,
-            json=expected_response_list_pull_requests,
-        )
-
-        generator = self._create_generator(self.current_tag, self.last_tag)
-        provider = GithubChangeNotesProvider(generator, self.current_tag, self.last_tag)
-        self.assertEquals(list(provider()), ['pull 1','pull 2','pull 3'])
-
-class TestReleaseNotesGenerator(unittest.TestCase):
-    
-    def setUp(self):
-        self.current_tag = 'prod/1.4'
-        self.last_tag = 'prod/1.3'
-        self.github_info = {
-            'github_owner': 'TestOwner',
-            'github_repo': 'TestRepo',
-            'github_username': 'TestUser',
-            'github_password': 'TestPass',
-        }
-
-    def test_init_without_last_tag(self):
-        github_info = self.github_info.copy()
-        generator = GithubReleaseNotesGenerator(github_info, self.current_tag)
-        self.assertEquals(generator.github_info, github_info)
-        self.assertEquals(generator.current_tag, self.current_tag)
-        self.assertEquals(generator.last_tag, None)
-        self.assertEquals(generator.change_notes.current_tag, self.current_tag)
-        self.assertEquals(generator.change_notes._last_tag, None)
-
-    def test_init_without_last_tag(self):
-        github_info = self.github_info.copy()
-        generator = GithubReleaseNotesGenerator(github_info, self.current_tag, self.last_tag)
-        self.assertEquals(generator.github_info, github_info)
-        self.assertEquals(generator.current_tag, self.current_tag)
-        self.assertEquals(generator.last_tag, self.last_tag)
-        self.assertEquals(generator.change_notes.current_tag, self.current_tag)
-        self.assertEquals(generator.change_notes._last_tag, self.last_tag)
-
